# swagger-website.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce Website API",
    "version": "1.0.0",
    "description": "API documentation for the E-Commerce Website"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/website/payments/webhook": {
      "post": {
        "summary": "Stripe webhook for payment events",
        "tags": [
          "Website Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook received successfully"
          }
        }
      }
    },
...
  },
...
}